Righted my wrongs
